@@ -1,3 +1,5 @@
+import java.text.SimpleDateFormat
+
 //
 // build.gradle in FtcRobotController
 //
@@ -6,11 +8,12 @@ apply plugin: 'com.android.library'
 android {
 
     defaultConfig {
-        minSdkVersion 19
+        minSdkVersion 23
         targetSdkVersion 28
+        buildConfigField "String", "BUILD_TIME", '"' + (new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ", Locale.ROOT).format(new Date())) + '"'
     }
 
-    compileSdkVersion 28
+    compileSdkVersion 29
 
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_7
@@ -19,6 +22,8 @@ android {
 }
 
 repositories {
+    maven { url = "https://dl.bintray.com/first-tech-challenge/ftcsdk/" }
+
     flatDir {
         dirs '../libs'
     }
